requirements.txt
# Core Dependencies
fastapi>=0.100
uvicorn[standard]>=0.24
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14
psycopg[binary]>=3.1
alembic>=1.12

# Authentication & Security
paseto>=2.0
argon2-cffi>=23.0

# AWS Integration (optional)
# Uncomment to enable AWS Secrets Manager support
# boto3==1.34.34
